+++ src/bin/node.rs
@@ -1,0 +1,77 @@
+mod click;
+
+use click::*;
+use clickhouse::Client;
+use dotenv::dotenv;
+use tracing_subscriber::EnvFilter;
+
+const PROJECT_ID: &str = "compact_indexer";
+
+fn main() {
+    openssl_probe::init_ssl_cert_env_vars();
+    dotenv().ok();
+
+    let args: Vec<String> = std::env::args().collect();
+    let home_dir = std::path::PathBuf::from(near_indexer::get_default_home());
+
+    let mut env_filter = EnvFilter::new(
+        "clickhouse=info,tokio_reactor=info,near=info,stats=info,telemetry=info,indexer=info,aggregated=info,compact_indexer=info",
+    );
+
+    if let Ok(rust_log) = std::env::var("RUST_LOG") {
+        if !rust_log.is_empty() {
+            for directive in rust_log.split(',').filter_map(|s| match s.parse() {
+                Ok(directive) => Some(directive),
+                Err(err) => {
+                    eprintln!("Ignoring directive `{}`: {}", s, err);
+                    None
+                }
+            }) {
+                env_filter = env_filter.add_directive(directive);
+            }
+        }
+    }
+
+    tracing_subscriber::fmt::Subscriber::builder()
+        .with_env_filter(env_filter)
+        .with_writer(std::io::stderr)
+        .init();
+
+    tracing::log::info!(target: PROJECT_ID, "Starting indexer",);
+
+    let command = args
+        .get(1)
+        .map(|arg| arg.as_str())
+        .expect("You need to provide a command: `init` or `run` as arg");
+
+    match command {
+        "run" => {
+            let client = establish_connection();
+            let indexer_config = near_indexer::IndexerConfig {
+                home_dir,
+                sync_mode: near_indexer::SyncModeEnum::FromInterruption,
+                await_for_node_synced: near_indexer::AwaitForNodeSyncedEnum::StreamWhileSyncing,
+                validate_genesis: false,
+            };
+            let sys = actix::System::new();
+            sys.block_on(async move {
+                let indexer = near_indexer::Indexer::new(indexer_config).unwrap();
+                let stream = indexer.streamer();
+                listen_blocks(stream, client).await;
+
+                actix::System::current().stop();
+            });
+            sys.run().unwrap();
+        }
+        _ => panic!("You have to pass `run` arg"),
+    }
+}
+
+async fn listen_blocks(
+    mut stream: tokio::sync::mpsc::Receiver<near_indexer::StreamerMessage>,
+    client: Client,
+) {
+    while let Some(streamer_message) = stream.recv().await {
+        extract_info(&client, streamer_message).await.unwrap();
+    }
+}
